removed constructor and refactored code into two methods
